Move: 컴포넌트 파일 하위 세부 폴더로 이동
